requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email-validator==2.1.0.post1
exceptiongroup==1.3.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
oauth2client==3.0.0
oauthlib==3.3.0
openai==1.61.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.23
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.1.0
Werkzeug==3.1.3
